Refactored cmd/client/main.go
