+++ internal/client/gclient.go
@@ -2,10 +2,15 @@ package client
 
 import (
 	"context"
+	"crypto/tls"
+	"log"
 
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/metadata"
 
+	"github.com/zklevsha/gophkeeper/internal/config"
 	"github.com/zklevsha/gophkeeper/internal/pb"
 )
 
@@ -16,7 +21,36 @@ type Gclient struct {
 }
 
 // NewGclient initializes new Gclient
-func NewGclient(conn *grpc.ClientConn) Gclient {
+func NewGclient(clientConfig config.ClientConfig, mstorage MemStorage) Gclient {
+
+
+	var conn *grpc.ClientConn
+	var err error
+	if clientConfig.UseTLS {
+		tlsConfig := &tls.Config{
+			InsecureSkipVerify: true,
+		}
+		conn, err = grpc.Dial(clientConfig.ServerAddress,
+			grpc.WithTransportCredentials(credentials.NewTLS(tlsConfig)),
+			grpc.WithUnaryInterceptor(getUnaryClientInterceptor(&mstorage)))
+	} else {
+		conn, err = grpc.Dial(clientConfig.ServerAddress,
+			grpc.WithTransportCredentials(insecure.NewCredentials()),
+			grpc.WithUnaryInterceptor(getUnaryClientInterceptor(&mstorage)))
+
+	}
+	if err != nil {
+		log.Fatalf("did not connect: %v", err)
+	}
+
+	defer func() {
+		err := conn.Close()
+		if err != nil {
+			log.Fatal(err)
+		}
+	}()
+
+
 	return Gclient{
 		Auth:  pb.NewAuthClient(conn),
 		Pdata: pb.NewPrivateDataClient(conn)}
@@ -29,8 +63,7 @@ var noAuth = map[string]bool{
 	"/Auth/GetToken": true,
 }
 
-// GetUnaryClientInterceptor returns a client interceptor
-func GetUnaryClientInterceptor(mstorage *MemStorage) grpc.UnaryClientInterceptor {
+func getUnaryClientInterceptor(mstorage *MemStorage) grpc.UnaryClientInterceptor {
 	return func(
 		ctx context.Context,
 		method string,
